+++ rviz_mesh_tools_plugins/include/rviz_mesh_tools_plugins/ClusterLabelPanel.hpp
@@ -50,11 +50,11 @@
 #ifndef CLUSTER_LABEL_PANEL_HPP
 #define CLUSTER_LABEL_PANEL_HPP
 
-#include <rviz_mesh_map_plugin/Types.hpp>
+#include <rviz_mesh_tools_plugins/Types.hpp>
 #include <QObject>
 #include <rviz_common/panel.hpp>
 #include <rviz_common/tool_manager.hpp>
-#include <rviz_mesh_map_plugin/ClusterLabelTool.hpp>
+#include <rviz_mesh_tools_plugins/ClusterLabelTool.hpp>
 #include <mesh_msgs/msg/mesh_face_cluster.hpp>
 
 // Forward declarations
@@ -66,7 +66,7 @@ namespace rviz
 class Tool;
 }
 
-namespace rviz_mesh_map_plugin
+namespace rviz_mesh_tools_plugins
 {
 /**
  * @class ClusterLabelPanel
@@ -144,6 +144,6 @@ protected:
   // ros::NodeHandle m_nodeHandle;
 };
 
-}  // end namespace rviz_mesh_map_plugin
+}  // end namespace rviz_mesh_tools_plugins
 
 #endif