+++ rviz_mesh_tools_plugins/include/rviz_mesh_tools_plugins/RVizMessageFilter.hpp
@@ -1,5 +1,5 @@
-#ifndef RVIZ_MESH_MAP_PLUGIN_RVIZ_MESSAGE_FILTER_HPP
-#define RVIZ_MESH_MAP_PLUGIN_RVIZ_MESSAGE_FILTER_HPP
+#ifndef RVIZ_MESH_TOOLS_PLUGINS_RVIZ_MESSAGE_FILTER_HPP
+#define RVIZ_MESH_TOOLS_PLUGINS_RVIZ_MESSAGE_FILTER_HPP
 
 #include <tf2_ros/message_filter.h>
 #include <rviz_common/transformation/frame_transformer.hpp>
@@ -15,4 +15,4 @@ using RVizMessageFilterPtr = std::shared_ptr<RVizMessageFilter<T> >;
 
 } // namespace tf2_ros
 
-#endif // RVIZ_MESH_MAP_PLUGIN_RVIZ_MESSAGE_FILTER_HPP
+#endif // RVIZ_MESH_TOOLS_PLUGINS_RVIZ_MESSAGE_FILTER_HPP