+++ rviz_mesh_tools_plugins/src/MeshVisual.cpp
@@ -50,7 +50,7 @@
  *    Jan Philipp Vogtherr <jvogtherr@uni-osnabrueck.de>
  */
 
-#include "rviz_mesh_map_plugin/MeshVisual.hpp"
+#include "rviz_mesh_tools_plugins/MeshVisual.hpp"
 
 #include <OgreSubEntity.h>
 #include <OgreRenderOperation.h>
@@ -64,7 +64,7 @@
 
 #include "rclcpp/rclcpp.hpp"
 
-namespace rviz_mesh_map_plugin
+namespace rviz_mesh_tools_plugins
 {
 Ogre::ColourValue getRainbowColor1(float value)
 {
@@ -107,7 +107,7 @@ MeshVisual::MeshVisual(rviz_common::DisplayContext* context, size_t displayID, s
   , m_texture_coords_enabled(false)
   , m_normalsScalingFactor(1)
 {
-  RCLCPP_INFO(rclcpp::get_logger("rviz_mesh_map_plugin"), "Creating MeshVisual %lu_TexturedMesh_%lu_%lu", m_prefix, m_postfix, m_random);
+  RCLCPP_INFO(rclcpp::get_logger("rviz_mesh_tools_plugins"), "Creating MeshVisual %lu_TexturedMesh_%lu_%lu", m_prefix, m_postfix, m_random);
 
   // get or create the scene node
   Ogre::SceneManager* sceneManager = m_displayContext->getSceneManager();
@@ -161,7 +161,7 @@ MeshVisual::MeshVisual(rviz_common::DisplayContext* context, size_t displayID, s
 
 MeshVisual::~MeshVisual()
 {
-  RCLCPP_INFO(rclcpp::get_logger("rviz_mesh_map_plugin"), "Destroying MeshVisual %lu_TexturedMesh_%lu_%lu", m_prefix, m_postfix, m_random);
+  RCLCPP_INFO(rclcpp::get_logger("rviz_mesh_tools_plugins"), "Destroying MeshVisual %lu_TexturedMesh_%lu_%lu", m_prefix, m_postfix, m_random);
 
   reset();
 
@@ -192,7 +192,7 @@ MeshVisual::~MeshVisual()
 
 void MeshVisual::reset()
 {
-  RCLCPP_INFO(rclcpp::get_logger("rviz_mesh_map_plugin"), "Resetting MeshVisual %lu_TexturedMesh_%lu_%lu", m_prefix, m_postfix, m_random);
+  RCLCPP_INFO(rclcpp::get_logger("rviz_mesh_tools_plugins"), "Resetting MeshVisual %lu_TexturedMesh_%lu_%lu", m_prefix, m_postfix, m_random);
 
   std::stringstream sstm;
 
@@ -557,7 +557,7 @@ void MeshVisual::enteringTriangleMeshWithVertexCosts(const Geometry& mesh, const
   float range = maxCost - minCost;
   if (range <= 0)
   {
-    RCLCPP_ERROR(rclcpp::get_logger("rviz_mesh_map_plugin"), "Illegal vertex cost limits!");
+    RCLCPP_ERROR(rclcpp::get_logger("rviz_mesh_tools_plugins"), "Illegal vertex cost limits!");
     return;
   }
 
@@ -654,7 +654,7 @@ void MeshVisual::enteringTexturedTriangleMesh(const Geometry& mesh, const vector
       // if the image was only default constructed, in which case its width will be 0
       if (m_images.size() < textureIndex + 1 || m_images[textureIndex].getWidth() == 0)
       {
-        RCLCPP_DEBUG(rclcpp::get_logger("rviz_mesh_map_plugin"), "Texture with index %u not loaded yet", textureIndex);
+        RCLCPP_DEBUG(rclcpp::get_logger("rviz_mesh_tools_plugins"), "Texture with index %u not loaded yet", textureIndex);
       }
       else
       {
@@ -783,7 +783,7 @@ bool MeshVisual::setGeometry(const Geometry& mesh)
   // check if there are enough vertices given
   if (mesh.vertices.size() < 3)
   {
-    RCLCPP_WARN(rclcpp::get_logger("rviz_mesh_map_plugin"), "Received not enough vertices, can't create mesh!");
+    RCLCPP_WARN(rclcpp::get_logger("rviz_mesh_tools_plugins"), "Received not enough vertices, can't create mesh!");
     return false;
   }
 
@@ -807,12 +807,12 @@ bool MeshVisual::setNormals(const vector<Normal>& normals)
   // check if there are vertex normals for each vertex
   if (normals.size() == m_geometry.vertices.size())
   {
-    RCLCPP_INFO(rclcpp::get_logger("rviz_mesh_map_plugin"), "Received %lu vertex normals.", normals.size());
+    RCLCPP_INFO(rclcpp::get_logger("rviz_mesh_tools_plugins"), "Received %lu vertex normals.", normals.size());
     m_vertex_normals_enabled = true;
   }
   else if (normals.size() > 0)
   {
-    RCLCPP_WARN(rclcpp::get_logger("rviz_mesh_map_plugin"), "Received not as much vertex normals as vertices, ignoring vertex normals!");
+    RCLCPP_WARN(rclcpp::get_logger("rviz_mesh_tools_plugins"), "Received not as much vertex normals as vertices, ignoring vertex normals!");
     return false;
   }
 
@@ -836,12 +836,12 @@ bool MeshVisual::setVertexColors(const vector<Color>& vertexColors)
   // check if there are vertex colors for each vertex
   if (vertexColors.size() == m_geometry.vertices.size())
   {
-    RCLCPP_INFO(rclcpp::get_logger("rviz_mesh_map_plugin"), "Received %lu vertex colors.", vertexColors.size());
+    RCLCPP_INFO(rclcpp::get_logger("rviz_mesh_tools_plugins"), "Received %lu vertex colors.", vertexColors.size());
     m_vertex_colors_enabled = true;
   }
   else
   {
-    RCLCPP_WARN(rclcpp::get_logger("rviz_mesh_map_plugin"), "Received not as much vertex colors as vertices, ignoring the vertex colors!");
+    RCLCPP_WARN(rclcpp::get_logger("rviz_mesh_tools_plugins"), "Received not as much vertex colors as vertices, ignoring the vertex colors!");
     return false;
   }
 
@@ -867,12 +867,12 @@ bool MeshVisual::setVertexCosts(const std::vector<float>& vertexCosts, int costC
   // check if there are vertex costs for each vertex
   if (vertexCosts.size() == m_geometry.vertices.size())
   {
-    RCLCPP_DEBUG(rclcpp::get_logger("rviz_mesh_map_plugin"), "Received %lu vertex costs.", vertexCosts.size());
+    RCLCPP_DEBUG(rclcpp::get_logger("rviz_mesh_tools_plugins"), "Received %lu vertex costs.", vertexCosts.size());
     m_vertex_costs_enabled = true;
   }
   else
   {
-    RCLCPP_WARN(rclcpp::get_logger("rviz_mesh_map_plugin"), "Received not as much vertex costs as vertices, ignoring the vertex costs!");
+    RCLCPP_WARN(rclcpp::get_logger("rviz_mesh_tools_plugins"), "Received not as much vertex costs as vertices, ignoring the vertex costs!");
     return false;
   }
 
@@ -896,12 +896,12 @@ bool MeshVisual::setVertexCosts(const std::vector<float>& vertexCosts, int costC
   // check if there are vertex costs for each vertex
   if (vertexCosts.size() == m_geometry.vertices.size())
   {
-    RCLCPP_DEBUG(rclcpp::get_logger("rviz_mesh_map_plugin"), "Received %lu vertex costs.", vertexCosts.size());
+    RCLCPP_DEBUG(rclcpp::get_logger("rviz_mesh_tools_plugins"), "Received %lu vertex costs.", vertexCosts.size());
     m_vertex_costs_enabled = true;
   }
   else
   {
-    RCLCPP_WARN(rclcpp::get_logger("rviz_mesh_map_plugin"), "Received not as much vertex costs as vertices, ignoring the vertex costs!");
+    RCLCPP_WARN(rclcpp::get_logger("rviz_mesh_tools_plugins"), "Received not as much vertex costs as vertices, ignoring the vertex costs!");
     return false;
   }
 
@@ -917,12 +917,12 @@ bool MeshVisual::setMaterials(const vector<Material>& materials, const vector<Te
   // check if there is a material index for each cluster
   if (materials.size() >= 0)
   {
-    RCLCPP_INFO(rclcpp::get_logger("rviz_mesh_map_plugin"), "Received %lu materials.", materials.size());
+    RCLCPP_INFO(rclcpp::get_logger("rviz_mesh_tools_plugins"), "Received %lu materials.", materials.size());
     m_materials_enabled = true;  // enable materials
   }
   else
   {
-    RCLCPP_WARN(rclcpp::get_logger("rviz_mesh_map_plugin"), "Received zero materials, ignoring materials!");
+    RCLCPP_WARN(rclcpp::get_logger("rviz_mesh_tools_plugins"), "Received zero materials, ignoring materials!");
     return false;
   }
 
@@ -930,13 +930,13 @@ bool MeshVisual::setMaterials(const vector<Material>& materials, const vector<Te
   // check if there are texture coords for each vertex
   if (texCoords.size() == m_geometry.vertices.size())
   {
-    RCLCPP_INFO(rclcpp::get_logger("rviz_mesh_map_plugin"), "Received %lu texture coords.", texCoords.size());
+    RCLCPP_INFO(rclcpp::get_logger("rviz_mesh_tools_plugins"), "Received %lu texture coords.", texCoords.size());
     m_texture_coords_enabled = true;  // enable texture coords
     m_textures_enabled = true;        // enable textures
   }
   else if (texCoords.size() > 0)
   {
-    RCLCPP_WARN(rclcpp::get_logger("rviz_mesh_map_plugin"), "Received not as much texture coords as vertices, ignoring texture coords!");
+    RCLCPP_WARN(rclcpp::get_logger("rviz_mesh_tools_plugins"), "Received not as much texture coords as vertices, ignoring texture coords!");
   }
 
   enteringTexturedTriangleMesh(m_geometry, materials, texCoords);
@@ -965,7 +965,7 @@ bool MeshVisual::addTexture(Texture& texture, uint32_t textureIndex)
   }
   else
   {
-    RCLCPP_WARN(rclcpp::get_logger("rviz_mesh_map_plugin"), "Can't load image into texture material, material does not exist!");
+    RCLCPP_WARN(rclcpp::get_logger("rviz_mesh_tools_plugins"), "Can't load image into texture material, material does not exist!");
     return false;
   }
 }
@@ -981,7 +981,7 @@ Ogre::PixelFormat MeshVisual::getOgrePixelFormatFromRosString(std::string encodi
     return Ogre::PF_BYTE_RGB;
   }
 
-  RCLCPP_WARN(rclcpp::get_logger("rviz_mesh_map_plugin"), "Unknown texture encoding! Using Ogre::PF_UNKNOWN");
+  RCLCPP_WARN(rclcpp::get_logger("rviz_mesh_tools_plugins"), "Unknown texture encoding! Using Ogre::PF_UNKNOWN");
   return Ogre::PF_UNKNOWN;
 }
 
@@ -1035,4 +1035,4 @@ void MeshVisual::setFrameOrientation(const Ogre::Quaternion& orientation)
   m_sceneNode->setOrientation(orientation);
 }
 
-}  // end namespace rviz_mesh_map_plugin
+}  // end namespace rviz_mesh_tools_plugins