+++ rviz_mesh_tools_plugins/CHANGELOG.rst
@@ -1,5 +1,5 @@
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Changelog for package rviz_mesh_map_plugin
+Changelog for package rviz_mesh_tools_plugins
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 1.1.0 (2021-10-27)
@@ -21,7 +21,7 @@ Changelog for package rviz_mesh_map_plugin
 * MeshDisplay now woring with MeshMap
 * added properties to MeshDisplay
 * added Map3D file dialog property
-* added textured mesh display for mesh msgs visualization to rviz_mesh_map_plugin
+* added textured mesh display for mesh msgs visualization to rviz_mesh_tools_plugins
 * combined mesh-plugin and map-plugin texture mesh visuals
 * fixed mpi error
 * resolved catkin lint problems